Refactor error handling and improve code readability in model inference nodes
- Enhanced error handling in `LoadTextEmbedderNode` and `TextEmbeddingNode` by using consistent formatting for validation errors.
- Improved code readability by restructuring the `execute` method in `LoadTextEmbedderNode` for better clarity.
- Added new variants to `LlmApiStyle` in `system_config.rs` for compatibility with Tencent multimodal APIs.
- Introduced utility functions in `openai_message_util.rs` for parsing and serializing OpenAI messages, enhancing modularity.
- Implemented streaming message parsing in `parse_chat_completions_sse_stream` and `parse_responses_sse_stream` for better handling of SSE responses.
- Added a new module `utils/mod.rs` to organize utility functions related to OpenAI message handling.
