+++ model_inference/src/nodes/llm_infer_node.rs
@@ -11,13 +11,20 @@ pub struct LLMInferNode {
 
 impl LLMInferNode {
     pub fn new(id: impl Into<String>, name: impl Into<String>) -> Self {
-        Self { id: id.into(), name: name.into() }
+        Self {
+            id: id.into(),
+            name: name.into(),
+        }
     }
 }
 
 impl Node for LLMInferNode {
-    fn id(&self) -> &str { &self.id }
-    fn name(&self) -> &str { &self.name }
+    fn id(&self) -> &str {
+        &self.id
+    }
+    fn name(&self) -> &str {
+        &self.name
+    }
     fn description(&self) -> Option<&str> {
         Some("使用LLModel引用对消息列表进行一次推理，返回模型回复")
     }
@@ -29,29 +36,53 @@ impl Node for LLMInferNode {
 
     node_output![port! { name = "response", ty = Vec(OpenAIMessage), desc = "LLM返回的消息列表" },];
 
-    fn execute(&mut self, inputs: HashMap<String, DataValue>) -> Result<HashMap<String, DataValue>> {
+    fn execute(
+        &mut self,
+        inputs: HashMap<String, DataValue>,
+    ) -> Result<HashMap<String, DataValue>> {
         self.validate_inputs(&inputs)?;
 
         let model = match inputs.get("llm_model") {
             Some(DataValue::LLModel(m)) => m.clone(),
-            _ => return Err(zihuan_core::error::Error::ValidationError("Missing required input: llm_model".to_string())),
+            _ => {
+                return Err(zihuan_core::error::Error::ValidationError(
+                    "Missing required input: llm_model".to_string(),
+                ));
+            }
         };
 
         let messages: Vec<OpenAIMessage> = match inputs.get("messages") {
-            Some(DataValue::Vec(_, items)) => items.iter().filter_map(|item| {
-                if let DataValue::OpenAIMessage(m) = item { Some(m.clone()) } else { None }
-            }).collect(),
-            _ => return Err(zihuan_core::error::Error::ValidationError("Missing required input: messages".to_string())),
+            Some(DataValue::Vec(_, items)) => items
+                .iter()
+                .filter_map(|item| {
+                    if let DataValue::OpenAIMessage(m) = item {
+                        Some(m.clone())
+                    } else {
+                        None
+                    }
+                })
+                .collect(),
+            _ => {
+                return Err(zihuan_core::error::Error::ValidationError(
+                    "Missing required input: messages".to_string(),
+                ));
+            }
         };
 
-        let param = InferenceParam { messages: &messages, tools: None };
+        let param = InferenceParam {
+            messages: &messages,
+            tools: None,
+        };
         let response_message = model.inference(&param);
 
         let mut outputs = HashMap::new();
-        outputs.insert("response".to_string(), DataValue::Vec(
-            Box::new(DataType::OpenAIMessage),
-            vec![DataValue::OpenAIMessage(response_message)],
-        ));
+        outputs.insert(
+            "response".to_string(),
+            DataValue::Vec(
+                Box::new(DataType::OpenAIMessage),
+                vec![DataValue::OpenAIMessage(response_message)],
+            ),
+        );
         self.validate_outputs(&outputs)?;
         Ok(outputs)
     }
