+++ model_inference/src/nodes/load_local_text_embedder_node.rs
@@ -13,15 +13,24 @@ pub struct LoadLocalTextEmbedderNode {
 
 impl LoadLocalTextEmbedderNode {
     pub fn new(id: impl Into<String>, name: impl Into<String>) -> Self {
-        Self { id: id.into(), name: name.into() }
+        Self {
+            id: id.into(),
+            name: name.into(),
+        }
     }
 }
 
 impl Node for LoadLocalTextEmbedderNode {
-    fn id(&self) -> &str { &self.id }
-    fn name(&self) -> &str { &self.name }
+    fn id(&self) -> &str {
+        &self.id
+    }
+    fn name(&self) -> &str {
+        &self.name
+    }
     fn description(&self) -> Option<&str> {
-        Some("从 models/text_embedding/<model_name> 加载本地 Candle embedding 模型，输出 EmbeddingModel 引用")
+        Some(
+            "从 models/text_embedding/<model_name> 加载本地 Candle embedding 模型，输出 EmbeddingModel 引用",
+        )
     }
 
     node_input![
@@ -32,17 +41,27 @@ impl Node for LoadLocalTextEmbedderNode {
         port! { name = "embedding_model", ty = EmbeddingModel, desc = "Embedding 模型引用" },
     ];
 
-    fn execute(&mut self, inputs: HashMap<String, DataValue>) -> Result<HashMap<String, DataValue>> {
+    fn execute(
+        &mut self,
+        inputs: HashMap<String, DataValue>,
+    ) -> Result<HashMap<String, DataValue>> {
         self.validate_inputs(&inputs)?;
 
         let model_name = match inputs.get("model_name") {
             Some(DataValue::String(value)) if !value.trim().is_empty() => value.trim().to_string(),
-            _ => return Err(Error::ValidationError("Missing required input: model_name".to_string())),
+            _ => {
+                return Err(Error::ValidationError(
+                    "Missing required input: model_name".to_string(),
+                ));
+            }
         };
 
         let model: Arc<dyn EmbeddingBase> = Arc::new(QueuedEmbeddingModel::new(model_name)?);
 
-        let outputs = HashMap::from([("embedding_model".to_string(), DataValue::EmbeddingModel(model))]);
+        let outputs = HashMap::from([(
+            "embedding_model".to_string(),
+            DataValue::EmbeddingModel(model),
+        )]);
         self.validate_outputs(&outputs)?;
         Ok(outputs)
     }
